requirements.txt
pytorch
torchvision
opencv-python
pandas
numpy